move IVerseSuggesting.ts
